# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  description?: string;
  sourceType: string;
  sourceRef?: string;
  status: string;
  languages: string[];
  /** @nullable */
  lastScanAt: string | null;
  createdAt: string;
}

export type ProjectInputSourceType = typeof ProjectInputSourceType[keyof typeof ProjectInputSourceType];


export const ProjectInputSourceType = {
  repository: 'repository',
  archive: 'archive',
  workspace: 'workspace',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  sourceType: ProjectInputSourceType;
  /** @minLength 1 */
  sourceRef: string;
  languages?: string[];
}

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Scan {
  id: number;
  projectId: number;
  projectName?: string;
  status: ScanStatus;
  stage: string;
  progress: number;
  startedAt: string;
  /** @nullable */
  completedAt: string | null;
  findingCount: number;
  /** @nullable */
  assuranceScore: number | null;
}

export type ScanInputMode = typeof ScanInputMode[keyof typeof ScanInputMode];


export const ScanInputMode = {
  autonomous: 'autonomous',
  assisted: 'assisted',
} as const;

export interface ScanInput {
  projectId: number;
  mode?: ScanInputMode;
}

export type ScanStageStatus = typeof ScanStageStatus[keyof typeof ScanStageStatus];


export const ScanStageStatus = {
  pending: 'pending',
  running: 'running',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface ScanStage {
  key: string;
  label: string;
  status: ScanStageStatus;
  detail: string;
  /** @nullable */
  durationMs?: number | null;
}

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type FindingStatus = typeof FindingStatus[keyof typeof FindingStatus];


export const FindingStatus = {
  open: 'open',
  confirmed: 'confirmed',
  patch_ready: 'patch_ready',
  verified: 'verified',
} as const;

export interface Finding {
  id: number;
  scanId: number;
  title: string;
  cwe: string;
  severity: FindingSeverity;
  file: string;
  line: number;
  function: string;
  status: FindingStatus;
  confidence: number;
  summary: string;
}

export interface ProofSummary {
  score: number;
  reproduced: boolean;
  patched: boolean;
  blocked: boolean;
  regressionPassed: boolean;
  staticClean: boolean;
  dynamicClean: boolean;
}

export type ScanDetail = Scan & {
  stages: ScanStage[];
  findings: Finding[];
  proof: ProofSummary;
};

export type EvidenceStatus = typeof EvidenceStatus[keyof typeof EvidenceStatus];


export const EvidenceStatus = {
  pass: 'pass',
  blocked: 'blocked',
  clean: 'clean',
  info: 'info',
} as const;

export interface Evidence {
  label: string;
  value: string;
  status: EvidenceStatus;
}

export interface Verification {
  exploitBefore: string;
  patchApplied: string;
  exploitAfter: string;
  regression: string;
  staticRescan: string;
  dynamicRetest: string;
  verdict: string;
}

export type FindingDetail = Finding & {
  rootCause: string;
  attackSurface: string;
  payload: string;
  evidence: Evidence[];
  beforeCode: string;
  afterCode: string;
  verification: Verification;
};

export type ActivityEventKind = typeof ActivityEventKind[keyof typeof ActivityEventKind];


export const ActivityEventKind = {
  scan: 'scan',
  finding: 'finding',
  verification: 'verification',
  system: 'system',
} as const;

export interface ActivityEvent {
  id: number;
  /** @nullable */
  scanId?: number | null;
  message: string;
  detail: string;
  kind: ActivityEventKind;
  createdAt: string;
}

export type DashboardSummarySeverity = {
  critical: number;
  high: number;
  medium: number;
  low: number;
};

export type DashboardSummaryAssuranceTrendItem = {
  label: string;
  score: number;
};

export interface DashboardSummary {
  projects: number;
  scans: number;
  findings: number;
  severity: DashboardSummarySeverity;
  latestScan: Scan | null;
  assuranceTrend: DashboardSummaryAssuranceTrendItem[];
}

export type ListScansParams = {
projectId?: number;
};

export type ListFindingsParams = {
scanId?: number;
};