# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryAssuranceTrendItem } from './dashboardSummaryAssuranceTrendItem';
import type { DashboardSummarySeverity } from './dashboardSummarySeverity';
import type { Scan } from './scan';

export interface DashboardSummary {
  projects: number;
  scans: number;
  findings: number;
  severity: DashboardSummarySeverity;
  latestScan: Scan | null;
  assuranceTrend: DashboardSummaryAssuranceTrendItem[];
}