# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverity } from './findingSeverity';
import type { FindingStatus } from './findingStatus';

export interface Finding {
  id: number;
  scanId: number;
  title: string;
  cwe: string;
  severity: FindingSeverity;
  file: string;
  line: number;
  function: string;
  status: FindingStatus;
  confidence: number;
  summary: string;
}