# lib/api-zod/src/generated/types/findingDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Evidence } from './evidence';
import type { Finding } from './finding';
import type { Verification } from './verification';

export type FindingDetail = Finding & {
  rootCause: string;
  attackSurface: string;
  payload: string;
  evidence: Evidence[];
  beforeCode: string;
  afterCode: string;
  verification: Verification;
};