# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectInputSourceType } from './projectInputSourceType';

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  sourceType: ProjectInputSourceType;
  /** @minLength 1 */
  sourceRef: string;
  languages?: string[];
}