# lib/api-zod/src/generated/types/proofSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProofSummary {
  score: number;
  reproduced: boolean;
  patched: boolean;
  blocked: boolean;
  regressionPassed: boolean;
  staticClean: boolean;
  dynamicClean: boolean;
}