# lib/api-zod/src/generated/types/scanDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Finding } from './finding';
import type { ProofSummary } from './proofSummary';
import type { Scan } from './scan';
import type { ScanStage } from './scanStage';

export type ScanDetail = Scan & {
  stages: ScanStage[];
  findings: Finding[];
  proof: ProofSummary;
};