# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanInputMode } from './scanInputMode';

export interface ScanInput {
  projectId: number;
  mode?: ScanInputMode;
}