# lib/api-zod/src/generated/types/verification.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Verification {
  exploitBefore: string;
  patchApplied: string;
  exploitAfter: string;
  regression: string;
  staticRescan: string;
  dynamicRetest: string;
  verdict: string;
}